--- src/main/java/com/kuit/healthmate/domain/HabitTime.java
@@ -1,4 +1,0 @@
-package com.kuit.healthmate.domain;
-
-public class HabitTime {
-}
