+++ src/main/java/com/kuit/healthmate/domain/habit/entity/Habit.java
@@ -1,9 +1,9 @@
-package com.kuit.healthmate.domain;
+package com.kuit.healthmate.domain.habit.entity;
 
+import com.kuit.healthmate.domain.User;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.ColumnDefault;
