+++ src/main/java/com/kuit/healthmate/domain/habit/entity/HabitTime.java
@@ -1,0 +1,4 @@
+package com.kuit.healthmate.domain.habit.entity;
+
+public class HabitTime {
+}
