+++ src/main/java/com/kuit/healthmate/domain/habit/repository/HabitRepository.java
@@ -1,6 +1,6 @@
-package com.kuit.healthmate.repository;
+package com.kuit.healthmate.domain.habit.repository;
 
-import com.kuit.healthmate.domain.Habit;
+import com.kuit.healthmate.domain.habit.entity.Habit;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
